Move BaseTransport into its own package in transports.basetransport
