+++ kasa/protocol.py
@@ -18,7 +18,7 @@ import hashlib
 import logging
 import struct
 from abc import ABC, abstractmethod
-from typing import Any, Callable, TypeVar, cast
+from typing import TYPE_CHECKING, Any, Callable, TypeVar, cast
 
 # When support for cpython older than 3.11 is dropped
 # async_timeout can be replaced with asyncio.timeout
@@ -30,6 +30,10 @@ _NO_RETRY_ERRORS = {errno.EHOSTDOWN, errno.EHOSTUNREACH, errno.ECONNREFUSED}
 _UNSIGNED_INT_NETWORK_ORDER = struct.Struct(">I")
 
 _T = TypeVar("_T")
+
+
+if TYPE_CHECKING:
+    from kasa.transports.basetransport import BaseTransport
 
 
 def redact_data(data: _T, redactors: dict[str, Callable[[Any], Any] | None]) -> _T:
@@ -73,49 +77,6 @@ def mask_mac(mac: str) -> str:
 def md5(payload: bytes) -> bytes:
     """Return the MD5 hash of the payload."""
     return hashlib.md5(payload).digest()  # noqa: S324
-
-
-class BaseTransport(ABC):
-    """Base class for all TP-Link protocol transports."""
-
-    DEFAULT_TIMEOUT = 5
-
-    def __init__(
-        self,
-        *,
-        config: DeviceConfig,
-    ) -> None:
-        """Create a protocol object."""
-        self._config = config
-        self._host = config.host
-        self._port = config.port_override or self.default_port
-        self._credentials = config.credentials
-        self._credentials_hash = config.credentials_hash
-        if not config.timeout:
-            config.timeout = self.DEFAULT_TIMEOUT
-        self._timeout = config.timeout
-
-    @property
-    @abstractmethod
-    def default_port(self) -> int:
-        """The default port for the transport."""
-
-    @property
-    @abstractmethod
-    def credentials_hash(self) -> str | None:
-        """The hashed credentials used by the transport."""
-
-    @abstractmethod
-    async def send(self, request: str) -> dict:
-        """Send a message to the device and return a response."""
-
-    @abstractmethod
-    async def close(self) -> None:
-        """Close the transport.  Abstract method to be overriden."""
-
-    @abstractmethod
-    async def reset(self) -> None:
-        """Reset internal state."""
 
 
 class BaseProtocol(ABC):
