+++ kasa/smartprotocol.py
@@ -12,7 +12,7 @@ import logging
 import time
 import uuid
 from pprint import pformat as pf
-from typing import Any, Callable
+from typing import TYPE_CHECKING, Any, Callable
 
 from .exceptions import (
     SMART_AUTHENTICATION_ERRORS,
@@ -26,7 +26,11 @@ from .exceptions import (
     _RetryableError,
 )
 from .json import dumps as json_dumps
-from .protocol import BaseProtocol, BaseTransport, mask_mac, md5, redact_data
+from .protocol import BaseProtocol, mask_mac, md5, redact_data
+
+if TYPE_CHECKING:
+    from kasa.transports.basetransport import BaseTransport
+
 
 _LOGGER = logging.getLogger(__name__)
