+++ kasa/transports/klaptransport.py
@@ -64,10 +64,10 @@ from kasa.httpclient import HttpClient
 from kasa.json import loads as json_loads
 from kasa.protocol import (
     DEFAULT_CREDENTIALS,
-    BaseTransport,
     get_default_credentials,
     md5,
 )
+from kasa.transports.basetransport import BaseTransport
 
 _LOGGER = logging.getLogger(__name__)
