+++ tests/fakeprotocol_smartcamera.py
@@ -5,7 +5,7 @@ from json import loads as json_loads
 
 from kasa import Credentials, DeviceConfig, SmartProtocol
 from kasa.experimental.smartcameraprotocol import SmartCameraProtocol
-from kasa.protocol import BaseTransport
+from kasa.transports.basetransport import BaseTransport
 
 from .fakeprotocol_smart import FakeSmartTransport
